% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHCL.R
\name{scHCL}
\alias{scHCL}
\title{Human Cell Landscape mapping
description}
\usage{
scHCL(scdata, numbers_plot = 3)
}
\arguments{
\item{scdata}{data.frame or matrix, col correspond to cells and rows correspond to genes}

\item{numbers_plot}{default is 3, number, it will return top "numbers_plot" records in the plot}
}
\description{
Human Cell Landscape mapping
description
}
