% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scHCL_gui.R
\name{scHCL_vis}
\alias{scHCL_vis}
\title{Visualize scHCL result in a web browser.}
\usage{
scHCL_vis(hcl_result)
}
\arguments{
\item{hcl_result}{an object generate by scHCL funtion}
}
\description{
Runs interactive \code{shiny} session of \code{scHCL}
}
